+++ EventManagement/src/main/java/com/example/eventmanagement/repository/EventRepository.java
@@ -44,7 +44,7 @@ public interface EventRepository extends JpaRepository<Event, Long> {
     @Query("SELECT COALESCE(SUM(tr.numberOfTickets), 0L) " +
            "FROM TicketReservation tr " +
            "WHERE tr.event.id = :eventId AND tr.bookingStatus = com.example.eventmanagement.enums.BookingStatus.CONFIRMED")
-    Long countConfirmedTicketsByEventId(@Param("eventId") Long eventId);
+    Integer countConfirmedTicketsByEventId(@Param("eventId") Long eventId);
 
     @Query("SELECT e, COUNT(t) as reservationCount " +
            "FROM Event e LEFT JOIN e.ticketReservations t " +
